Change App.js to class component
