+++ src/App.js
@@ -1,12 +1,13 @@
 /* eslint-disable */
+import React, { Component } from 'react';
 import Calculator from './components/Calculator';
 
-function App() {
-  return (
-    <div className="App">
-      <Calculator />
-    </div>
-  );
+export default class App extends Component {
+  render() {
+    return (
+      <div className="App">
+        <Calculator />
+      </div>
+    );
+  }
 }
-
-export default App;
